model_server/requirements.txt
#TOOD: pin versions
fastapi
sentence-transformers
torch
uvicorn
gliner
transformers
pyyaml
accelerate
# guard inference packages
optimum-intel
openvino
psutil
